Move Opeartion.hpp into an operations folder
