+++ raster-cli/include/operations/Operation.hpp
